Restauro de funcionalidade

+++ src/main/SpaceManager.java
@@ -30,11 +30,11 @@ public class SpaceManager extends Thread {
 	public int getAvailableSpace() {
 		return totalSpace;
 	}
-	
+
 	public ArrayList<Chunk> getStoredChunks(){
 		return this.storedChunks;
 	}
-	
+
 	public void setAvailableSpace(int totalSpace) {
 		this.totalSpace = totalSpace;
 	}
@@ -69,6 +69,9 @@ public class SpaceManager extends Thread {
 							chunkToRemove = nextChunk;
 					}
 				}
+				if (chunkToRemove == null || chunkToRemove.getActualReplicationDegree() <= 1) continue;
+
+				System.out.println("Deleting chunk " + chunkToRemove.getFileId() + " " + chunkToRemove.getChunkNumber());
 
 				File toRemove = new File("chunks" + File.separator + chunkToRemove.getFileId() + File.separator + chunkToRemove.getChunkNumber());
 				toRemove.delete();
